createOrUpdateNodeInfoCRD
